CRAFTERCMS-570 fixed issues
